requirements.txt
langchain>=1.0.0
langchain-openai>=0.1.0
langchain-core>=1.0.0
numpy>=1.24.0
ipykernel>=6.25.0
python-dotenv>=1.0.0
